EN-23: Add NUnit test framework
