+++ global.json
@@ -1,0 +1,6 @@
+{
+    "projects": [
+        "src",
+        "test"
+    ]
+}
