+++ src/BalanceApi/Controllers/BaseController.cs
@@ -1,0 +1,19 @@
+using Microsoft.AspNetCore.Mvc;
+using System.Net;
+using System;
+
+namespace BalanceApi.Controllers
+{
+    public abstract class BaseController : Controller
+    {
+
+        protected IActionResult ForException(Exception ex)
+        {
+            return InternalServerError(ex.Message);
+        }
+        protected IActionResult InternalServerError(object payload)
+        {
+            return StatusCode((int)HttpStatusCode.InternalServerError, payload);
+        }
+    }
+}
