+++ src/BalanceApi/Model/Domain/AppSettings.cs
