+++ src/BalanceApi/Model/Domain/Result.cs
