+++ src/BalanceApi/Model/Domain/TransactionType.cs
