+++ src/BalanceApi/Validators/TransactionTypeValidator.cs
