+++ src/BalanceApi/appsettings.json
@@ -1,0 +1,14 @@
+{
+  "Logging": {
+    "IncludeScopes": false,
+    "LogLevel": {
+      "Default": "Information",
+      "System": "Information",
+      "Microsoft": "Information"
+    }
+  },
+
+  "AppSettings": {
+    "ConnectionString": "Server=localhost;Database=account_balance;Uid=root;Pwd=r00t;SslMode=None"
+  }
+}
