+++ src/BalanceApi/project.json
@@ -2,7 +2,8 @@
   "version": "1.0.0-*",
   "buildOptions": {
     "debugType": "portable",
-    "emitEntryPoint": true
+    "emitEntryPoint": true,
+    "copyToOutput" : ["appsettings.json"]
   },
   "dependencies": {
       "Microsoft.AspNetCore.Server.Kestrel": "1.1.0",
