+++ test/BalanceApi.Tests/project.json
@@ -1,0 +1,26 @@
+{
+  "version": "1.0.0-*",
+  "testRunner": "nunit",
+  "dependencies": {
+      "NUnit" : "3.5.0",
+      "dotnet-test-nunit" : "3.4.0-beta-3",
+      "BalanceApi" : {
+        "target" : "project"
+      },
+
+      "Microsoft.DotNet.InternalAbstractions" : "1.0.0"
+          
+  },
+
+  "frameworks": {
+    "netcoreapp1.1": {
+      "dependencies": {
+        "Microsoft.NETCore.App": {
+          "type": "platform",
+          "version": "1.1.0"
+        }
+      },
+      "imports": [ "dotnet45" ]
+    }
+  }
+}
